More bin work! :D

+++ common/mekanism/common/block/BlockMachine.java
@@ -148,7 +148,7 @@ public class BlockMachine extends BlockContainer implements ISpecialBounds
     public void onBlockPlacedBy(World world, int x, int y, int z, EntityLivingBase entityliving, ItemStack itemstack)
     {
     	TileEntityBasicBlock tileEntity = (TileEntityBasicBlock)world.getBlockTileEntity(x, y, z);
-        int side = MathHelper.floor_double((double)(entityliving.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+        int side = MathHelper.floor_double((entityliving.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
         int height = Math.round(entityliving.rotationPitch);
         int change = 3;
